test(opensearchtransport): tag the zero-allocation assertions !integration
TestClassify_ZeroAlloc and TestNewRequestEventZeroAlloc use
testing.AllocsPerRun, a process-wide allocation differential. Neither
source file carried a build tag, so both compiled into every
configuration including -tags integration,core, sharing a binary with the
live-cluster tests whose transports poll node stats and cluster health in
the background. That is what flaked TestClassify_ZeroAlloc in CI.

Move each test, unchanged, into a //go:build !integration file so the
differential only ever runs where nothing else allocates concurrently.
The two tests need separate files because their sources sit in different
packages: classify_extra_test.go is opensearchtransport_test and
observer_response_internal_test.go is opensearchtransport. The remaining
tests in both source files stay untagged; only the differentials need the
constraint.

Add TestCloseReapsBackgroundPollers as the durable guard: it asserts via
the goroutine dump that New starts the node-stats and cluster-health
pollers and that Close reaps both. Verified to fail when Close is stubbed
to skip its context cancellation.
