+++ opensearchtransport/observer_response_internal_test.go
@@ -384,26 +384,3 @@ func TestTracingHooksLifecycle(t *testing.T) {
 	require.True(t, obs.respCtxOK, "request-scoped context reaches OnRequestResponse")
 	require.Equal(t, 1, obs.attempts)
 }
-
-// TestNewRequestEventZeroAlloc guards that building the RequestEvent snapshot
-// fired at observers allocates nothing: all fields are scalars or strings copied
-// from the pre-captured streamResult (Host from the connection's cached
-// hostPort), so the value stays on the stack.
-func TestNewRequestEventZeroAlloc(t *testing.T) {
-	req := &http.Request{
-		Method: http.MethodGet,
-		URL:    &url.URL{Scheme: "http", Host: "node-1:9200", Path: "/idx/_search"},
-	}
-	sr := streamResult{
-		escapedPath: "/idx/_search",
-		routeName:   "search",
-		index:       "idx",
-		poolName:    "search",
-		hostPort:    "http://node-1:9200",
-	}
-	allocs := testing.AllocsPerRun(100, func() {
-		ev := newRequestEvent(req, sr)
-		_ = ev
-	})
-	require.Zero(t, allocs, "newRequestEvent must not allocate")
-}
